requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
pydantic==2.3.0
pydantic[email]==2.3.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
httpx==0.24.1
python-multipart==0.0.6